tests: split config processing and arg validation tests
