+++ tests/unit_tests/config_test.py
@@ -176,44 +176,3 @@ def test_process_config_dict_strategy_invalid(caplog):
     assert args.strategy == Strategy.Spatial
     assert caplog.records[0].message == "Invalid strategy 'an-invalid-strategy', reverting to default value Strategy.Spatial"
 
-
-def test_process_config_dict_blend_invalid(caplog):
-    args = CausalArgs()
-    config_dict = {"causal": {"distribution": {"blend": 20}}}
-
-    with pytest.raises(ReXTomlError):
-        process_config_dict(config_dict, args)
-        assert caplog.records[0].message == "Invalid value '20': must be between 0.0 and 1.0"
-
-
-def test_process_config_dict_permitted_overlap_invalid(caplog):
-    args = CausalArgs()
-    config_dict = {"explanation": {"multi": {"permitted_overlap": -5}}}
-
-    with pytest.raises(ReXTomlError):
-        process_config_dict(config_dict, args)
-        assert caplog.records[0].message == "Invalid value '-5': must be between 0.0 and 1.0"
-
-
-def test_process_config_dict_iters_invalid(caplog):
-    args = CausalArgs()
-    config_dict = {"causal": {"iters": 0}}
-    with pytest.raises(ReXTomlError):
-        process_config_dict(config_dict, args)
-        assert caplog.records[0].message == "Invalid value '0': must be more than 0.0"
-
-
-def test_process_config_dict_raw_invalid(caplog):
-    args = CausalArgs()
-    config_dict = {"rex": {"visual": {"raw": 100 }}}
-    with pytest.raises(ReXTomlError):
-        process_config_dict(config_dict, args)
-        assert caplog.records[0].message == "Invalid value '100': must be boolean"
-
-
-def test_process_config_dict_multi_style_invalid(caplog):
-    args = CausalArgs()
-    config_dict = {"rex": {"visual": {"multi_style": "an-invalid-style"}}}
-    with pytest.raises(ReXTomlError):
-        process_config_dict(config_dict, args)
-        assert caplog.records[0].message == "Invalid value 'an-invalid-style' for multi_style, must be 'composite' or 'separate'"
