+++ tests/unit_tests/preprocessing_test.py
@@ -2,8 +2,8 @@ import pytest
 from rex_xai.explanation import (
     predict_target,
     try_preprocess,
-    validate_args,
 )
+from rex_xai.config import validate_args
 
 def test_preprocess_nii_notimplemented(args, model_shape, cpu_device, caplog):
     args.path = "tests/test_data/dog.nii"
@@ -19,8 +19,3 @@ def test_predict_target(data, prediction_func):
     assert target.classification == 207
     assert target.confidence == pytest.approx(0.253237, abs=2.5e-6)
 
-
-def test_validate_args(args):
-    args.path = None  #  type: ignore
-    with pytest.raises(FileNotFoundError):
-        validate_args(args)
